add missing changes
