+++ docs/source/utilities.rst
@@ -14,48 +14,48 @@ set_host_device_count
 Inference Utilities
 ===================
 
-.. automodule:: numpyro.infer_util
+.. automodule:: numpyro.infer.util
 
 predictive
 ----------
-.. autofunction:: numpyro.infer_util.predictive
+.. autofunction:: numpyro.infer.util.predictive
 
 log_density
 -----------
-.. autofunction:: numpyro.infer_util.log_density
+.. autofunction:: numpyro.infer.util.log_density
 
 transform_fn
 ------------
-.. autofunction:: numpyro.infer_util.transform_fn
+.. autofunction:: numpyro.infer.util.transform_fn
 
 constrain_fn
 ------------
-.. autofunction:: numpyro.infer_util.constrain_fn
+.. autofunction:: numpyro.infer.util.constrain_fn
 
 potential_energy
 ----------------
-.. autofunction:: numpyro.infer_util.potential_energy
+.. autofunction:: numpyro.infer.util.potential_energy
 
 log_likelihood
 --------------
-.. autofunction:: numpyro.infer_util.log_likelihood
+.. autofunction:: numpyro.infer.util.log_likelihood
 
 find_valid_initial_params
 -------------------------
-.. autofunction:: numpyro.infer_util.find_valid_initial_params
+.. autofunction:: numpyro.infer.util.find_valid_initial_params
 
 init_to_median
 --------------
-.. autofunction:: numpyro.infer_util.init_to_median
+.. autofunction:: numpyro.infer.util.init_to_median
 
 init_to_prior
 -------------
-.. autofunction:: numpyro.infer_util.init_to_prior
+.. autofunction:: numpyro.infer.util.init_to_prior
 
 init_to_uniform
 ---------------
-.. autofunction:: numpyro.infer_util.init_to_uniform
+.. autofunction:: numpyro.infer.util.init_to_uniform
 
 init_to_feasible
 ----------------
-.. autofunction:: numpyro.infer_util.init_to_feasible
+.. autofunction:: numpyro.infer.util.init_to_feasible
