+++ examples/neutra.py
@@ -18,8 +18,7 @@ from numpyro.diagnostics import summary
 import numpyro.distributions as dist
 from numpyro.distributions import constraints
 from numpyro.infer import MCMC, NUTS, SVI, elbo
-from numpyro.infer.hmc_util import initialize_model
-from numpyro.infer_util import transformed_potential_energy
+from numpyro.infer.util import initialize_model, transformed_potential_energy
 
 # TODO: remove when the issue https://github.com/google/jax/issues/939 is fixed upstream
 # The behaviour when training guide under fast math mode is unstable.
