+++ numpyro/__init__.py
@@ -1,4 +1,4 @@
-from numpyro import compat, diagnostics, distributions, handlers, infer, infer_util, optim
+from numpyro import compat, diagnostics, distributions, handlers, infer, optim
 import numpyro.patch  # noqa: F401
 from numpyro.primitives import module, param, plate, sample
 from numpyro.util import set_host_device_count, set_platform
@@ -14,7 +14,6 @@ __all__ = [
     'distributions',
     'handlers',
     'infer',
-    'infer_util',
     'module',
     'optim',
     'param',
