+++ numpyro/infer/__init__.py
@@ -1,8 +1,22 @@
 from numpyro.infer.mcmc import HMC, MCMC, NUTS
 from numpyro.infer.svi import SVI, elbo
+from numpyro.infer.util import (
+    init_to_feasible,
+    init_to_median,
+    init_to_prior,
+    init_to_uniform,
+    log_likelihood,
+    predictive
+)
 
 __all__ = [
     'elbo',
+    'init_to_feasible',
+    'init_to_median',
+    'init_to_prior',
+    'init_to_uniform',
+    'predictive',
+    'log_likelihood',
     'HMC',
     'MCMC',
     'NUTS',
