+++ test/test_hmc_util.py
@@ -1,5 +1,4 @@
 from collections import namedtuple
-from functools import partial
 import logging
 import os
 
@@ -10,10 +9,7 @@ import pytest
 from jax import device_put, disable_jit, grad, jit, random, tree_map
 import jax.numpy as np
 
-import numpyro
-from numpyro import handlers
 import numpyro.distributions as dist
-from numpyro.distributions.transforms import biject_to
 from numpyro.infer.hmc_util import (
     AdaptWindow,
     _is_iterative_turning,
@@ -23,20 +19,10 @@ from numpyro.infer.hmc_util import (
     consensus,
     dual_averaging,
     find_reasonable_step_size,
-    initialize_model,
     parametric_draws,
-    potential_energy,
     velocity_verlet,
     warmup_adapter,
     welford_covariance
-)
-from numpyro.infer_util import (
-    constrain_fn,
-    init_to_feasible,
-    init_to_median,
-    init_to_prior,
-    init_to_uniform,
-    transform_fn
 )
 from numpyro.util import control_flow_prims_disabled, fori_loop, optional
 
@@ -403,97 +389,6 @@ def test_build_tree(step_size):
     # for small step_size, assert that it should take a while to meet the terminate condition
     if step_size < 0.1:
         assert tree.num_proposals > 10
-
-
-def test_model_with_transformed_distribution():
-    x_prior = dist.HalfNormal(2)
-    y_prior = dist.LogNormal(scale=3.)  # transformed distribution
-
-    def model():
-        numpyro.sample('x', x_prior)
-        numpyro.sample('y', y_prior)
-
-    params = {'x': np.array(-5.), 'y': np.array(7.)}
-    model = handlers.seed(model, random.PRNGKey(0))
-    inv_transforms = {'x': biject_to(x_prior.support), 'y': biject_to(y_prior.support)}
-    expected_samples = partial(transform_fn, inv_transforms)(params)
-    expected_potential_energy = (
-        - x_prior.log_prob(expected_samples['x']) -
-        y_prior.log_prob(expected_samples['y']) -
-        inv_transforms['x'].log_abs_det_jacobian(params['x'], expected_samples['x']) -
-        inv_transforms['y'].log_abs_det_jacobian(params['y'], expected_samples['y'])
-    )
-
-    base_inv_transforms = {'x': biject_to(x_prior.support), 'y': biject_to(y_prior.base_dist.support)}
-    actual_samples = constrain_fn(
-        handlers.seed(model, random.PRNGKey(0)), (), {}, base_inv_transforms, params)
-    actual_potential_energy = potential_energy(model, (), {}, base_inv_transforms, params)
-
-    assert_allclose(expected_samples['x'], actual_samples['x'])
-    assert_allclose(expected_samples['y'], actual_samples['y'])
-    assert_allclose(actual_potential_energy, expected_potential_energy)
-
-
-@pytest.mark.parametrize('init_strategy', [
-    init_to_feasible,
-    partial(init_to_median, num_samples=2),
-    init_to_prior,
-    init_to_uniform,
-])
-def test_initialize_model_change_point(init_strategy):
-    def model(data):
-        alpha = 1 / np.mean(data)
-        lambda1 = numpyro.sample('lambda1', dist.Exponential(alpha))
-        lambda2 = numpyro.sample('lambda2', dist.Exponential(alpha))
-        tau = numpyro.sample('tau', dist.Uniform(0, 1))
-        lambda12 = np.where(np.arange(len(data)) < tau * len(data), lambda1, lambda2)
-        numpyro.sample('obs', dist.Poisson(lambda12), obs=data)
-
-    count_data = np.array([
-        13,  24,   8,  24,   7,  35,  14,  11,  15,  11,  22,  22,  11,  57,
-        11,  19,  29,   6,  19,  12,  22,  12,  18,  72,  32,   9,   7,  13,
-        19,  23,  27,  20,   6,  17,  13,  10,  14,   6,  16,  15,   7,   2,
-        15,  15,  19,  70,  49,   7,  53,  22,  21,  31,  19,  11,  18,  20,
-        12,  35,  17,  23,  17,   4,   2,  31,  30,  13,  27,   0,  39,  37,
-        5,  14,  13,  22,
-    ])
-
-    rngs = random.split(random.PRNGKey(1), 2)
-    init_params, _, _ = initialize_model(rngs, model, count_data,
-                                         init_strategy=init_strategy)
-    for i in range(2):
-        init_params_i, _, _ = initialize_model(rngs[i], model, count_data,
-                                               init_strategy=init_strategy)
-        for name, p in init_params.items():
-            # XXX: the result is equal if we disable fast-math-mode
-            assert_allclose(p[i], init_params_i[name], atol=1e-6)
-
-
-@pytest.mark.parametrize('init_strategy', [
-    init_to_feasible,
-    partial(init_to_median, num_samples=2),
-    init_to_prior,
-    init_to_uniform,
-])
-def test_initialize_model_dirichlet_categorical(init_strategy):
-    def model(data):
-        concentration = np.array([1.0, 1.0, 1.0])
-        p_latent = numpyro.sample('p_latent', dist.Dirichlet(concentration))
-        numpyro.sample('obs', dist.Categorical(p_latent), obs=data)
-        return p_latent
-
-    true_probs = np.array([0.1, 0.6, 0.3])
-    data = dist.Categorical(true_probs).sample(random.PRNGKey(1), (2000,))
-
-    rngs = random.split(random.PRNGKey(1), 2)
-    init_params, _, _ = initialize_model(rngs, model, data,
-                                         init_strategy=init_strategy)
-    for i in range(2):
-        init_params_i, _, _ = initialize_model(rngs[i], model, data,
-                                               init_strategy=init_strategy)
-        for name, p in init_params.items():
-            # XXX: the result is equal if we disable fast-math-mode
-            assert_allclose(p[i], init_params_i[name], atol=1e-6)
 
 
 # TODO: raise this warning issue upstream, the issue is at this line
